refactor login page to include components of sign up and sign in forms instead of directly on page
